pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-triage_agent_env"
version = "0.1.0"
description = "TriageAgent environment for OpenEnv — enterprise IT ticket triage via multi-turn tool use"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "numpy>=1.24",
    "anthropic>=0.40",
    "openai>=1.0",
    "matplotlib>=3.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "triage_agent_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["triage_agent_env", "triage_agent_env.server"]
package-dir = { "triage_agent_env" = ".", "triage_agent_env.server" = "server" }
